app/requirements.txt
aniso8601==9.0.1
attrs==21.2.0
autopep8==1.6.0
blueprint==3.4.2
click==8.0.3
Flask==2.0.2
Flask-PyMongo==2.3.0
flask-restx==0.5.1
gunicorn==20.1.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.3.2
MarkupSafe==2.0.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
PyJWT==2.3.0
pymongo==4.0.1
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
toml==0.10.2
Werkzeug==2.0.2
